chore(tonic): Move ConnectError out of transport
